+++ crates/cubecl-runtime/src/memory_management/memory_pool/base.rs
@@ -9,7 +9,7 @@ use crate::{
 pub(crate) struct Slice {
     pub storage: StorageHandle,
     pub handle: SliceHandle,
-    pub padding: usize,
+    pub padding: u64,
 }
 
 impl Slice {
@@ -17,7 +17,7 @@ impl Slice {
         self.handle.is_free()
     }
 
-    pub(crate) fn effective_size(&self) -> usize {
+    pub(crate) fn effective_size(&self) -> u64 {
         self.storage.size() + self.padding
     }
 
@@ -26,7 +26,7 @@ impl Slice {
     }
 }
 
-pub(crate) fn calculate_padding(size: usize, buffer_alignment: usize) -> usize {
+pub(crate) fn calculate_padding(size: u64, buffer_alignment: u64) -> u64 {
     let remainder = size % buffer_alignment;
     if remainder != 0 {
         buffer_alignment - remainder
@@ -36,19 +36,18 @@ pub(crate) fn calculate_padding(size: usize, buffer_alignment: usize) -> usize {
 }
 
 pub trait MemoryPool {
-    fn max_alloc_size(&self) -> usize;
+    fn max_alloc_size(&self) -> u64;
 
     fn get(&self, binding: &SliceBinding) -> Option<&StorageHandle>;
 
     fn reserve<Storage: ComputeStorage>(
         &mut self,
         storage: &mut Storage,
-        size: usize,
+        size: u64,
         locked: Option<&MemoryLock>,
     ) -> SliceHandle;
 
-    fn alloc<Storage: ComputeStorage>(&mut self, storage: &mut Storage, size: usize)
-        -> SliceHandle;
+    fn alloc<Storage: ComputeStorage>(&mut self, storage: &mut Storage, size: u64) -> SliceHandle;
 
     fn get_memory_usage(&self) -> MemoryUsage;
